VirtioScsi: split to a separate file
