+++ pkg/frontend/scsi_test.go
@@ -1,0 +1,57 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (c) 2022 Dell Inc, or its subsidiaries.
+
+// Package frontend implememnts the FrontEnd APIs (host facing) of the storage Server
+package frontend
+
+import (
+	"testing"
+)
+
+func TestFrontEnd_CreateVirtioScsiController(t *testing.T) {
+
+}
+
+func TestFrontEnd_DeleteVirtioScsiController(t *testing.T) {
+
+}
+
+func TestFrontEnd_UpdateVirtioScsiController(t *testing.T) {
+
+}
+
+func TestFrontEnd_ListVirtioScsiControllers(t *testing.T) {
+
+}
+
+func TestFrontEnd_GetVirtioScsiController(t *testing.T) {
+
+}
+
+func TestFrontEnd_VirtioScsiControllerStats(t *testing.T) {
+
+}
+
+func TestFrontEnd_CreateVirtioScsiLun(t *testing.T) {
+
+}
+
+func TestFrontEnd_DeleteVirtioScsiLun(t *testing.T) {
+
+}
+
+func TestFrontEnd_UpdateVirtioScsiLun(t *testing.T) {
+
+}
+
+func TestFrontEnd_ListVirtioScsiLuns(t *testing.T) {
+
+}
+
+func TestFrontEnd_GetVirtioScsiLun(t *testing.T) {
+
+}
+
+func TestFrontEnd_VirtioScsiLunStats(t *testing.T) {
+
+}
